requirements.txt
requests==2.31.0
gTTS==2.5.1
SpeechRecognition==3.10.3
pygame==2.5.2
pydub==0.25.1
ffmpeg-python==0.2.0
wikipedia==1.4.0
python-dotenv==1.0.1
cryptography
sentence-transformers
torch
rapidfuzz
nltk
google-generativeai>=0.3.0